+++ sources/typescript/src/kaltura-clients/index.ts
@@ -1,3 +1,0 @@
-export * from './kaltura-browser-http-client';
-export * from './kaltura-client-base';
-export * from './kaltura-http-client-base';